Add test cases and fix command
