+++ .automation/test/go_modules/go_modules_bad/golang_bad_01.go
@@ -1,3 +1,10 @@
+package main
+
+import (
+	"github.com/go-playground/validator/v10"
+	"github.com/labstack/echo/v4"
+)
+
 if len(in) == 0 {
   return "", fmt.Errorf("Input is empty")
 }
