+++ .automation/test/go_modules/go_modules_good/golang_good_01.go
@@ -1,7 +1,31 @@
 package main
 
-import "fmt"
+import (
+	"github.com/go-playground/validator/v10"
+	"github.com/labstack/echo/v4"
+)
+
+type FooType struct {
+	Var *string `validate:"alphanum"`
+}
+
+func post(echo.Context) (err error) {
+	validate := validator.New()
+
+	if err = validate.Struct(FooType{Var: nil}); err != nil {
+		return
+	}
+
+	return
+}
+
+func server() (e *echo.Echo) {
+	e = echo.New()
+	e.POST("/", post)
+	return
+}
 
 func main() {
-	fmt.Println("hello world")
+	e := server()
+	e.Logger.Fatal(e.Start(":8000"))
 }
